Enhance: make Mudlet respect icons on dialog buttons system setting
At least on X11 based OSes it is possible for the Window Manager to change
the styling/images used for icons on native dialog boxes - or even hide
them.

This hides/shows the icons on dialogues (currently only the "connection
preferences") so that they match the system settings that the Qt generated
dialogues already respect.  There is not a QApplication::getSetting(option)
means to do this - instead we have to see whether the buttons on a Qt
generated dialogue using a predefined button role has an icon and use it's
presence as an indication.

This commit also overcomes the problem that the icons would previously
have shown as the "baked-in" icons in the "Welcome message" even if they
were not present on the buttons themselves.

In the future it will be desirable to also use QIcon::fromTheme(...) to
furnish the icons for other dialogues and also respect user/system settings
for showing/hiding icons on dialogue buttons where we add buttons ourselves
to such QDialogButtonBox - from past investigations this is particularly
going to apply to the Pack Manager; Module Manager and Package Exporter
dialogue as they current have an irregular set of buttons that I plan to
"tidy-up" by moving more of them into the button-box on their dialogues.

Also:
* delete now unused icon:
  * /src/icons/list-add_small.png
* replace with a "better" icon on the Connect Profile dialog (but used
  elsewhere also and left unchanged there):
  * /src/icons/edit-delete-shred.png ==> /src/icons/edit-delete.png
  copy profile button:
  * /src/icons/list-add_small.png ==> /src/icons/edit-copy.png
  new profile button:
  * /src/icons/list-add_small.png ==> /src/icons/document-new.png

Removed:
* empty #if defined(INCLUDE_UPDATER) code block from mudlet.cpp

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>