Cast to correct data type to avoid memory corruption (#1742)
This fixes a memory corruption on Windows when compiling Mudlet with MinGW 5.3. It might have been a problem on other platforms as well, but it became apparent on Windows with #1176. The memory corruption is triggered in the linked issue by denying the telnet subnegotiation for ATCP. The cast from char to int lead to a change at a negative index in a bool array which happened to modify the internal state of zlib which in turn lead to a segfault or faulty state (which stalled).