Enhance: add option to allow all Key-bindings that match key-press to run (#1656)
* Enhance: add option to allow all Key-bindings that match key-press to run

I had previously spotted that the process that checks for a key-binding
match on each key-press stops after the first match is found and run. This
is, IMHO, counter-intuitive given Mudlet's behaviour with other
"duplicated" items which it permits to exist - i.e. two triggers set to
match on the same text or two timers with the same interval; it is even
more suspect when one considers that it is possible to create a temporary
key-binding, which although active, may never get run if there happens to
be a permanent one with the same key and modifiers combination.  Actually
it is also possible that the reverse is true in that a temporary key
binding - which is not visible in the GUI editor - can prevent a permanent
one that is shown there from activating - with no visible indication that
this is occurring!

Anyhow, this commit adds a per-profile option on the "Input Line" tab of
the profile preferences dialogue to check every key-binding for a match
instead of stopping on the first match.  To maintain backwards
compatibility with scripts/packages/modules that depend on the previous
behaviour this option is not enabled by default.