Enhance: add manual override control for icons on menus
I found that my "system" setting was not being presented to Mudlet in the
value held returned by:
QCoreApplication::testAttribute(Qt::AA_DontShowIconsInMenus)
even after a reboot - so I felt it was necessary to provide a manual
override {though it defaults to not doing that!} The setting is stored and
is global.

Because we need the setting before the GUI is created - and because that
also applies to the forthcoming GUI locale/language setting - I have split
the mudlet::readSettings(const QSettings&) into two so that some of the
settings can be read before setupUi(...) is run in the mudlet constructor.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>