BugFix: try and restore reported key binding recent issues
This should allow using the Shift modifier with the spacebar to not be
different to how it was in 3.9.0 - but it does mean that that combination
is also now no longer usable as a key-binding.

This should allow using the shift and/or control modifier with the
backspace key as it was in 3.9.0 - but again it does mean that those
combinations are also now no longer usable as key-binding(s).

I also spotted a bug that meant that plain back-tab (or rather shift-tab)
was not processed correctly because the Shift that will probably accompany
it was preventing the code for plain back-tab from being reached.

These changes should fix Issue: #1751 .

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>