BugFix: copy single line text without leading spaces (#1774)
This fixes the recently introduced regression where selecting only part of
a line of text caused the insertion of spaces corresponding to the offset
of the starting character was introduce before the first character.  While
this WAS the intended action when more than one line of text is involved
it is NOT appropriate when the selection is only a single line.  This
revision has also been ported to the HTML copy case but give that that is
less likely to be used only for a part of just one line it was probably
less noticeable.

Also:
* Removed two redundant nesting of a none-slot method inside a slot wrapper
  - it is not needed and a slot type method can be invoked perfectly well
  with a direct call:
  * (void) TTextEdit::copySelectionToClipboard()
  * (void) TTextEdit::copySelectionToClipboardHTML()

* Optimise a pair of for(...) loops in the same area.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>