BugFix: restore focus to main TConsole after mouse click in other ones
This is less than perfect because it cannot totally identify what the
particular TConsole instance is associated with - to help me to debug it
I needed to add some extra code that assigns values to the
(QString) TConsole::mConsoleName to cover the miniConsoles, buffers, user
windows and Editor error window.

As well as attempting to close issue #1783 I think the addition of the
setFocusPolicy(Qt::NoFocus) to the mudlet::openWindow(...) method will also
cure #1769.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>