revise: unify Linux/FreeBSD/macOs keep-alive settings
A small change to the ctelnet.h file is needed to enable compilation on
FreeBSD to also use the (void) ctelnet::setKeepAlive(int) - which also does
not need the:
setsockopt(socketHandle, SOL_SOCKET, SO_KEEPALIVE, &on, sizeof(on))
call because it has defaulted to having that option turned-on for a
long-time.

Note that in the previous MacOs code it is likely that the line containing
"TCP_KEEPALIVE" is a noop because I suspect it is not defined (so will
be equivalent to 0 which is not defined to any option as far as I can
tell).

This commit restores all the settings to the defaults I believe except for
the delay before the first keep-alive is sent - which should be two minutes
now for the Windows case and possibly 75 + 120 = 195 (3 minutes 15 seconds)
depending on how long it takes for the socket to be opened in the other
OSs.

Hopefully the FIRST keep-alive will do it's work and keep the connection
open.  MUDLET will now close the socket after 10 x 75 = 12.5 minutes after
the first keep-alive gets sent and NO response is received from the other
end.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>