unify duplicate strings for i18n (#1804)
#### Brief overview of PR changes/additions
Removing the changing part like `(n of 5)` from translation comments of otherwise identical strings to archieve improved Crowdin integration.

#### Motivation for adding to Mudlet
There were comments like `This string is also used programmatically ensure all instances are the same, (1 of 5)` etc. but it seems like Crowdin can automatically ensure this. 
Compare the string in the same file with comment `This string is used in 2 places, ensure they match!` which is only listed once for translation in Crowdin, not twice, but indicates both line numbers as context:
![image](https://user-images.githubusercontent.com/117238/42607163-d2d6803a-8581-11e8-9d4a-e5d9d22aab55.png)

So to have the same effect here, we need to remove the changing part (n of 5) from source: 
![image](https://user-images.githubusercontent.com/117238/42607184-e1a6572a-8581-11e8-8bd9-17c1562af0a2.png)

Also some typographical correction to ensure understandability of the two ideas mentioned in the comment.

#### Other info (issues closed, discussion etc)
Then again, the comment is thereby obsolete and should be adjusted for more relevance to the actual contents of the string. This is left for another PR.

What is more, the surrounding < and > brackets will break Crowdin integration a bit, as Crowdin will assume this to be html code and hide it from translators. This somewhat breaks easy translatability. A possible solution may be along the same way that actual html code is handled e.g. around line 234: `QStringLiteral("<%1>").arg(tr("Text in brackets"))` Again, this is left for another PR.