BugFix: include a fail-back icon for the auto-saved profile in Con. Dialog
"filesave" is not present as a themed icon in some Distributions and
without a fall-back icon contained within the Mudlet resources no icon
appears at all - making it impossible to pick out the auto-saved entry.

This commit changes it to the "document-save" icon and includes the
corresponding Oxygen (so GPLed) 48x48 icon to provide the fall-back.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>