BugFix: fix faulty log options for new profiles or old profile save files
If a profile save file was from an older version it may not have an
`logFileNameFormat` attribute for the `<Host>` element this would clear the
`(QString) Host::mLogFileNameFormat` member away from the default coded
value. This commit now only changes the setting away from the default if
a setting is present in a saved file.

Additionally, the old default value for the DateTime specification string
used for some log file format types did not match any of the settings in
the "Profile Preference" control for this - with the result that a new
profile would show **no** setting when the preferences dialog was opened.
This commit updates a faulty DateTime setting by changing any "hh" read in
it to "HH" which then makes the rest of the code work as intended.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>