Revise: fixup a couple of loose ends in preparation for merge
Remove a debugging output line that is not useful now and will be spammy.

Modernise a triplet of QObject::connect(...) calls that will otherwise
clash harder when merged into development after "upgrade-to-qt5-connect"
PR has also been merged into the main branch.

NOTE: If merging this PR into development after the
"upgrade-to-qt5-connect" PR it will be necessary to resolve the referred to
merge clash so that the development branch's:
* (void) dlgTriggerEditor::slot_set_pattern_type_color(int type)
gets replaced by/split into:
* (QColor) dlgTriggerEditor::itemTypeColor(const int type)
* (void) dlgTriggerEditor::setupPatternControls(const int type,
                                              dlgTriggerPatternEdit* pItem)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>