Bugfixes: fix a previously spotted HTML error and an operational one
It had been previously noted that there was a use of bold html tags inside
a block of test that was already emboldened so that the inner ones did not
have any effect.

Fix a coding error when the initial use of
(void) dlgColorTrigger::setupBasicButtons(bool, bool isFirstTime = false)
was not called with the second defaulted argument supplied with a
non-default argument to include some run-once code.

Also:
* add some explanation/help text to the dlgColorTrigger dialog.
* revise the text explaining the formula (232 + grey scale value 0..23)
used for colours from the 24 grey-scale part of the 256 colour range.
* add some tool-tips to parts of the dlgColorTrigger dialog.
* add a generic static method to the mudlet class to provide a consistent
and re-usable "HTML" wrapper around text - which will be particularly
useful for tool-tip generation.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>