BugFix: attempt to fix UTF-8 handling of MSDP
Correct the cTelnet::processTelnetCommand(const string&) to correctly
measure the length of the msdp data in bytes (appropriate for UTF-8 encoded
data) and not QChars (appropriate for UTF-16 encoded QStrings) so that the
correct size is passed to (void) TLuaInterpreter::msdp2Lua(char *, int) -
this was changed to a const char * for the first argument to show the
intent to not modify the passed array, and because it is a const char * it
was also appropriate to change the length to the correct size_t type to use
when indexing into it.

Revise (void) TLuaInterpreter::msdp2Lua(const char *, size_t) to process
the supplied array as a QByteArray and NOT a QString - as it is processing
a single BYTE at a time and not a QChar at a time.  The conversion to a
QString is only done when the processed data is passed to
(void) TLuaInterpreter::setMSDPTable(QString&, const QString&) .

Some other clean up operations were done for MSDP - including unifying the
mix-mash of `69` and the `const char MSDP = 69` to the consistent
`const char OPT_MSDP = 69` which more closely matches other constants used
for other OOB protocols in the code.

Some raw strings were revised to be placed in `QLatin1String` wrappers when
used to initialise or compare against `QString` and some others removed
from `QLatin1Char` wrappers when they are now added or compared to
`QByteArrays`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>