BugFix: add more revisions to handle Mud Server encoding for OOB protocols
Rearrange logic to an `if` in (void) Host::send(...) to avoid an empty
true branch.

Remove (void) Host::sendRaw(QString) as it is not really required.

Add current version/build to MXP `<VERSION>` response - it was previously
fixed at 2.0 when it should be set to the actual version.

Add "[ WARN ] -" message if trying to send non-ASCII characters to a MUD
Server that is set to use ASCII encoding.

Change a local variable `_h` in several places where it is used to assemble
a std::string as using a variable with a name beginning with `_` possibly
not wise in a C++ environment where name mangling is used...

Removed some `#defined` constants from `TLuaInterpreter.cpp` and replace
them with some equivalent `const char`s defined in `cTelnet.h`.

Clean up the above mentioned `const char`s and use them in more places
where previously the direct value was placed as a raw literal.

Added handling of the Mud Server encoding of data in, modernised the code
style of, and gated the use of:
* `TLuaInterpreter::sendATCP(...)`
* `TLuaInterpreter::sendGMCP(...)`
* `TLuaInterpreter::sendTelnetChannel102(...)`
so that they only operate if the corresponding protocol *has* been
negotiated, this needed adding some previously missing
`enableXxxxx = false` entries to the processing of Telnet commands. Also
modernised the use of and updated it to handle the Mud Server encoding:
* `TLuaInterpreter::sendMSDP(...)`

Revise `(bool) cTelnet::socketOutRaw(string&)` to use
`(bool) QAbstractSocket::isValid()` instead of
`(bool) QAbstractSocket::isWritable()`, which seems a better fit and is
documented for use with Sockets, whereas the latter is inherited from the
`QIODevice` class. Also revise the code using QAbstractSocket::write(...)
to correctly handle a *partial* write of the data concerned, the prior code
did not take account of the need to resend only unsent bytes on incomplete
writes.

Also revise the variable type used to index into the bool[256] arrays of
flags in:
* cTelnet::myOptionState
* cTelnet::hisOptionState
* cTelnet::announcedState
* cTelnet::heAnnouncedState
* cTelnet::triedToEnable
although the last one is buggy as the bits are never set in that member and
the remainder are not correctly modified for those Telnet options that we
use within Mud's Telnet code - which means that Telnet sub-option 5 - STATUS
information will be WRONG!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>