Revise: fixes needed after revise and retesting
I found there are some non-printing Unicode code-points that it is sensible
to report - especially as they cannot currently be shown {the FitzPatrick
Emoji modifiers that give a skin colour tint to human related Emojis},
however as they are not on the BMP a revision was needed to:
`inline QString TTextEdit::convertWhiteSpaceToVisual(...)`

Some undefined code sequencing (using multiple operator++ calls in the
multiple operators of a method meant the order in which they were used
was arbitrary) and produced errors when done in the wrong order.

Added per profile option to enable (disabled by default) this analyser as
requested.

Added an additional line to the information table which gives the decimal
escape codes that would be needed to specify the individual bytes that
make up a non-ASCII grapheme in Lua for characters that are not printable
ASCII characters.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>