Initial QTextToSpeech additions (#1188)
Add lua functions to control an installed Text-to-Speech engine. All profiles share the same engine and queue their text in a common queue to avoid the cost of cooperation.

Currently, only the OS locale (language) can be used for speech output. Additionally, the queueing acts up if used too quick back to back (in the region of milliseconds).