From commit to edbee-lib:
Corrects placement of autocomplete list when editor isn't scrolled to the very top.
Updates tooltip placement to only display to the left if there is enough room left, and not enough room right. If there isn't room in either direction, it's shown right so there isn't confusion about whether the tooltip for that entry actually exists.