Bugfix/Enhance:  Suppress debug output from short interval timers
Adds a QDateTime spinbox control (only showing h:mm:ss:zzz) to the last
page of the profile preferences that allows the user to set an interval
for which any Timer less than this for the currently active Host which will
NOT report the "Lua OK" line in green - so the timer will not fill up with
spammy text so much. As this is to *temporarily* disable some debug output
this setting is NOT saved between sessions.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>