Enhance: allow a profile prefences dialog per Host
Some time back the code-base was revised to only allow one Profile
Preferences to be active at a time - this was so multiple clicks on the
menu item or toolbar button did not make loads of copies of the form /
dialogue.  However there are some advantages to being able to see the
settings for multiple profiles side-by-side.

This commit should provide this capability.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>