Enhance allow a profile preference dialog per host (#1908)
Some time back the code-base was revised to only allow one Profile
Preferences to be active at a time - this was so multiple clicks on the
menu item or toolbar button did not make loads of copies of the form /
dialogue.  However there are some advantages to being able to see the
settings for multiple profiles side-by-side.

To stop the common settings being confused/getting out of sync if more
than one profile is active and has the preference dialogue open saving them
by closing such a dialogue/form will now update the common settings
currently defined for that dialogue.

CARE SHOULD BE TAKEN THAT ANY OTHER "APPLICATION-WIDE" SETTING IS ALSO
HANDLED IN THE SAME WAY SO THAT CHANGING THE SETTING AND SAVING IT FROM
ONE PROFILE'S PREFERENCES DIALOGUE ARE PROPAGATED TO ANY OTHER OPEN
PREFERENCE DIALOGUE.

As I needed to ensure that several application wide values are updated in
any open profile preferences dialogues when the settings are saved in a
particular one I found it was worthwhile to start the process of moving the
enabling of the "Full Screen View" option from being controlled via a
sentinel file kept in the Mudlet base directory into the `QSetting` system
that is used for other application wide settings.  At present the existing
file will also be created as needed but I would suggest that support for it
is phased out between the future Mudlet 4.0 and 4.1 versions...

Also, in several places where multiple uses are made of mudlet::self() it
seemed sensible to taking a local copy of the returned (effectively
constant) pointer value (as `pMudlet`).

In passing I identified the correct `enum` values to use to control the
display of "WhiteSpace" in the Edbee widgets so have replaced the values
`0` and `1` with their correct symbols:
* `edbee::TextEditorConfig::ShowWhitespaces`
* `edbee::TextEditorConfig::HideWhitespaces`

instead.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>