Revise: allow display of only the partially selected part of the line
Previously this PR would show the entire first line that contained the
start of the selected text in the analysis - this commit revises things
so that only the selected part is shown. However the (1-based) indexes into
the QChar (QString, UTF16BE, C++ core view) and UTF-8 (Lua view) of the
selection are reworked so that they are still correct for a partially
selected line.

The connect(...) call has also been updated to the Qt5 style. Note that
this feature uses the QAction::hovered() signal which can cause problems
when used if running inside a debugger such as GDB from Qt Creator.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>