Repair: correct merge errors
Some HTML/Rich text entries in the dlgProfilePreferences constructor got
reverted to a form with an unnecessary QStringLiteral wrapper containing
a commonly used set of outer tags.

There were some entries that had got duplicated in the mudle.qrc XML file
which qmake (or the resource compiler) is unhappy about. By editing the
file by hand to sort the entries it was possible to find and clean up the
rogue entries. Unfortunately the Qt IDE does not attempt to keep the list
of files sorted and doing it manually after adding anything is the only
way to ensure that merges do not otherwise mess things up!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>