BugFix: handle Qt auto-accelerator creation on main tabBar
On some OS platforms (not macOs) the texts shown on the tabBar which shows
when more than one profile is loaded automagically gains a unique
accelerator for each different tab.  However this also modifies the text
returned by QTabBar::tabText(int) and, at least on my FreeBSB system the
displayed added underline is migrated into the text to appear as a '&' in
the text.  This BREAKS our codebase as that expects the tabText to be the
profile name and uses it as such.  This commit corrects that by storing and
using the profile name as held in the QTabBar tabs' data member instead so
that if the text gets changed the correct profile name is still accessible
to our code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>