Revise: improve updateqm.pri file
Improves the script in the updateqm.pri file that tries to identify what
the command to run the Lua interpreter is so that it can handle the
alternative `lua51` that FreeBSD uses for the (now a non-default) Lua 5.1
interpreter.

Added an advisory message to ignore the error message that the Windows
process (using `where` in `cmd.exe`) generates when the first one or two
tries to find the lua interpreter use the wrong file name.

Also tidied-up some poorly nested/indented qmake procedures which made it
hard to visually determine what was being attempted.

Added the `-compress` argument to the calls to `lrelease` in the qmake
project file so that the invocation produces the same effect as the CMake
one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>