Revise: change warning to accommodate gap between it and thing being warned
The previous text put the warning in the right sort of place on my
atypical Windows setup but that is not how the default Windows setup will
show it as it is not then at a point where it makes much sense as written.
This commit rephrases it so that the warning does fit in in either case
(my setup or the normal one).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>