Changed Geyser.Gauge:setValue to use text label
Geyser gauges have three layers, with one specifically for the text of the gauge. Geyser.Gauge:setText and Geyser.Gauge:setColor both used the text label for text they were given, but setValue still used the old style of writing to the front and back labels. This change updates setValue to follow the new method.