Fix missing file issue
Now properly handles missing files/folders as needed as it initializes. Also modified find_prompt to send a blank command, so a prompt will hopefully be shown right away, and added a timer so if no prompt is seen, it times out and gives a message instructing the user what to do next.