Revise: properly distinguish between two pairs of two languages
Using "English" to describe "American English" introduces some confusion
when "English (British)" is also listed. Similarly calling a language
"Chinese" when it actually means "Simplified Chinese" which is what is
generally used in China can be confusing when Taiwan uses "Traditional
Chinese" - this is actually a point of contention in that part of the
World and it is not wise to mix up what each setting is supposed to mean or
to imply a difference importance to different variants for the same
language.

Note that I have changed to use square brackets `[`...`]` around the
Country or other disambiguation term here because `(`...`)` are already
used for only partially translated languages and it looks messy to use
the same type of brackets twice in the same item when them imply completely
different things (%-age translated and variant)!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>