Revise: fix '<' being treated as HTML + move WS text comments into tr(...)
In the new TTextEdit::convertWhitespaceToVisual(...) {renamed from
"convertWhiteSpaceToVisual" !) I've moved the comments that describe each
"code" for each white space (for which there are thus little or no visible
means to tell them apart otherwise) characters into the translation tr(...)
call so that it is available to translators.  For most languages it will be
best not to attempt to translated these codes but for ideographic
lanaguages where each grapheme can be a complete word it is possible that
a complete explaination will be of comparible space on screen as the "code"
that explains what each white-space character is so is worth providing for
those languages only!

Also fixes a bug where the less-than character is seen as the code in Qt
Rich-Text/HTML that opens a formatting tag and is not displayed as a '<'.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>