Revise: some changes provoked by peer review
Changes the text of some of the lua run-time error messages for sending
ATCP and GMCP messages.

Limit the issuing in the main console of a warning about the use of
characters that cannot be sent with the currently selected Mud Server
encoding to only one per profile session - unless the encoding is changed.

Remove some effectless code that was copy-pasted from another class that
did need it.

Correct a couple of spelling errors in comments.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>