Fix handling of non-ASCII by OutOfBand protocols (MSDP/GMPC/ATCP) (#1852)
Correct the cTelnet::processTelnetCommand(const string&) to correctly
measure the length of the msdp data in bytes (appropriate for UTF-8 encoded
data) and not QChars (appropriate for UTF-16 encoded QStrings) so that the
correct size is passed to (void) TLuaInterpreter::msdp2Lua(char *, int) -
this was changed to a const char * for the first argument to show the
intent to not modify the passed array, and because it is a const char * it
was also appropriate to change the length to the correct size_t type to use
when indexing into it.

Revise (void) TLuaInterpreter::msdp2Lua(const char *, size_t) to process
the supplied array as a QByteArray and NOT a QString - as it is processing
a single BYTE at a time and not a QChar at a time.  The conversion to a
QString is only done when the processed data is passed to
(void) TLuaInterpreter::setMSDPTable(QString&, const QString&) .

Some other clean up operations were done for MSDP - including unifying the
mix-mash of `69` and the `const char MSDP = 69` to the consistent
`const char OPT_MSDP = 69` which more closely matches other constants used
for other OOB protocols in the code.

Some raw strings were revised to be placed in `QLatin1String` wrappers when
used to initialise or compare against `QString` and some others removed
from `QLatin1Char` wrappers when they are now added or compared to
`QByteArrays`.

Rearrange logic to an `if` in (void) Host::send(...) to avoid an empty
true branch.

Remove (void) Host::sendRaw(QString) as it is not really required.

Add current version/build to MXP `<VERSION>` response - it was previously
fixed at 2.0 when it should be set to the actual version.

Add "[ WARN ] -" message if trying to send non-ASCII characters to a MUD
Server that is set to use ASCII encoding.

Change a local variable `_h` in several places where it is used to assemble
a std::string as using a variable with a name beginning with `_` possibly
not wise in a C++ environment where name mangling is used...

Removed some `#defined` constants from `TLuaInterpreter.cpp` and replace
them with some equivalent `const char`s defined in `cTelnet.h`.

Clean up the above mentioned `const char`s and use them in more places
where previously the direct value was placed as a raw literal.

Added handling of the Mud Server encoding of data in, modernised the code
style of, and gated the use of:
* `TLuaInterpreter::sendATCP(...)`
* `TLuaInterpreter::sendGMCP(...)`
* `TLuaInterpreter::sendTelnetChannel102(...)`
so that they only operate if the corresponding protocol *has* been
negotiated, this needed adding some previously missing
`enableXxxxx = false` entries to the processing of Telnet commands. Also
modernised the use of and updated it to handle the Mud Server encoding:
* `TLuaInterpreter::sendMSDP(...)`

Revise `(bool) cTelnet::socketOutRaw(string&)` to use
`(bool) QAbstractSocket::isValid()` instead of
`(bool) QAbstractSocket::isWritable()`, which seems a better fit and is
documented for use with Sockets, whereas the latter is inherited from the
`QIODevice` class. Also revise the code using QAbstractSocket::write(...)
to correctly handle a *partial* write of the data concerned, the prior code
did not take account of the need to resend only unsent bytes on incomplete
writes.

Also revise the variable type used to index into the bool[256] arrays of
flags in:
* cTelnet::myOptionState
* cTelnet::hisOptionState
* cTelnet::announcedState
* cTelnet::heAnnouncedState
* cTelnet::triedToEnable
although the last one is buggy as the bits are never set in that member and
the remainder are not correctly modified for those Telnet options that we
use within Mud's Telnet code - which means that Telnet sub-option 5 - STATUS
information will be WRONG!

I initially used the version of QAbstractSocket::write(...) which had
only a single argument but that terminates writing bytes when it reaches
an ASCII NUL which put the method into an endless loop.  As some data that
gets sent out will contain embedded NUL characters (possibly as part of a
Telnet sub-option negotiation) this was causing fatal hangs, e.g. when
sending a TTYPE where the TNSB_IS (Telnet Sub-Option "IS") code is used.

Limit the issuing in the main console of a warning about the use of
characters that cannot be sent with the currently selected Mud Server
encoding to only one per profile session - unless the encoding is changed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>