BugFix: cure Coverity reported error
The code analysis that we have done by Coverity recently reported a dead
(unreachable) code issue that I introduced whilst trying to "improve" the
Telnet status reporting sub-option (5) - as a side effect of #1852.
Basically I was comparing a `size_t` loop variable (which ran through all
the values between 0 to 255 inclusive) to a `const char` which should have
had the value of 240 (I also had a coding error and was using a value of
250 but the same principle applied!)  Due, I think, to the intricacies of
`char` promotion to `int` a `char` value greater than 127 gets treated as a
negative value (as a `signed char`) rather than the `unsigned char` I was
expecting. I hope to fix that with this commit which makes the conversion
go via the `unsigned char` type which should make it work...

If successful, this will close Issue #2103.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>