Revise: dead code removal for main window
Remove some unused QActions defined in the ./ui/main_window.ui form:
* dactionForum_2
* dactionIRC_2
* actionTest1
* actionAsdf

Removed the `windowTitle` and windowIcon definitions in the main_window.ui
form as they are wrong and overwritten in the mudlet constructor anyway.

Remove some unused QActions defined in the mudlet constructor but never
placed into the application's windows:
* mactionConnect
* mactionAlias
* mactionTimers
* mactionButtons
* mactionScripts
* mactionKeys
* mactionMapper
* mactionHelp
* mactionOptions
* mactionMultiView
* mactionAbout

Moved the statusTip texts for items on the menu bar to the corresponding
toolTip texts because we have not had a status bar for a long time and thus
the texts were never shown - this did cause the replacement of one toolTip
text (for the QAction to open the editor)! Also converted all toolTip texts
in main_window.ui to be in a rich-text form (with initial capital letter
and full stop at the end) with the addition of HTML paragraph tags to wrap
the text.

A further one mactionCloseProfile has been left but has been commented out
because I have potentially a plan to add/restore a "Close" profile action
to the Game menu on the main menubar.

Some menubar QActions were renamed to match the form used by others:
* actionShow_Map         ==> dactionShowMap
* actionLive_Help_Chat   ==> dactionLiveHelpChat
* actionPackage_manager  ==> dactionPackageManager
* actionModule_manager   ==> dactionModuleManager
* actionPackage_Exporter ==> dactionPackageExporter

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>